// app/interfaces/MainComponent/IBlogsectionComponent.ts
export interface IBlogsection {
    _id: string;
    id?: Number;
    title?: string;
    image?: string;
    heading1?: string;
    heading2?: string;
    heading3?: string;
    heading4?: string;
    heading5?: string;
    heading6?: string;
    heading7?: string;
    heading8?: string;
    heading9?: string;
    heading10?: string;
    description1?: string;
    description2?: string;
    description3?: string;
    description4?: string;
    description5?: string;
    description6?: string;
    description7?: string;
    description8?: string;
    description9?: string;
    description10?: string;
  
}

export interface IBlogsectionInput {
    id?: Number;
    title?: string;
    image?: string;
    heading1?: string;
    heading2?: string;
    heading3?: string;
    heading4?: string;
    heading5?: string;
    heading6?: string;
    heading7?: string;
    heading8?: string;
    heading9?: string;
    heading10?: string;
    description1?: string;
    description2?: string;
    description3?: string;
    description4?: string;
    description5?: string;
    description6?: string;
    description7?: string;
    description8?: string;
    description9?: string;
    description10?: string;
}